Removed unnecessary casts (STR #2308).
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@7982 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ test/line_style.cxx
@@ -58,10 +58,10 @@ void test_box::draw() {
   dashes[3] = char(sliders[8]->value());
   dashes[4] = 0;
   fl_line_style(
-    (long)(choice[0]->mvalue()->argument()) +
-    (long)(choice[1]->mvalue()->argument()) +
-    (long)(choice[2]->mvalue()->argument()),
-    (long)(sliders[3]->value()), // width
+    choice[0]->mvalue()->argument() +
+    choice[1]->mvalue()->argument() +
+    choice[2]->mvalue()->argument(),
+    long(sliders[3]->value()),		// width
     dashes);
 
   // draw the defined fl_rect and fl_vertex first and then
